Move OidcConfigLookup logic into DefaultOidcConfigurationService
OidcConfigLookup was a standalone @Component whose sole purpose was
to dispatch OIDC config lookups to KinoticSystemService,
OrganizationService, or ApplicationService based on the auth scope
type. Fold it into DefaultOidcConfigurationService as
getConfigsForScope(authScopeType, authScopeId) and expose it on
the OidcConfigurationService interface. Delete OidcConfigLookup.

OrganizationService and ApplicationService are injected @Lazy to
break the circular dependency (ApplicationService depends on
OidcConfigurationService for getOidcConfigurations, and now
OidcConfigurationService dispatches back to ApplicationService).

IamSecurityService now injects OidcConfigurationService directly.
