+++ kinotic-domain/src/main/java/org/kinotic/os/internal/api/services/iam/DefaultOidcConfigurationService.java
@@ -5,10 +5,15 @@ import co.elastic.clients.elasticsearch.core.MgetRequest;
 import co.elastic.clients.elasticsearch.core.mget.MultiGetOperation;
 import org.apache.commons.lang3.Validate;
 import org.kinotic.core.api.security.ParticipantContext;
+import org.kinotic.os.api.model.iam.AuthScopeType;
 import org.kinotic.os.api.model.iam.OidcConfiguration;
+import org.kinotic.os.api.services.ApplicationService;
+import org.kinotic.os.api.services.KinoticSystemService;
+import org.kinotic.os.api.services.OrganizationService;
 import org.kinotic.os.api.services.iam.OidcConfigurationService;
 import org.kinotic.os.internal.api.services.AbstractCrudService;
 import org.kinotic.os.internal.api.services.CrudServiceTemplate;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Component;
 
 import java.util.Date;
@@ -19,10 +24,37 @@ import java.util.concurrent.CompletableFuture;
 @Component
 public class DefaultOidcConfigurationService extends AbstractCrudService<OidcConfiguration> implements OidcConfigurationService {
 
+    private final KinoticSystemService kinoticSystemService;
+    private final OrganizationService organizationService;
+    private final ApplicationService applicationService;
+
     public DefaultOidcConfigurationService(CrudServiceTemplate crudServiceTemplate,
                                            ElasticsearchAsyncClient esAsyncClient,
-                                           ParticipantContext participantContext) {
+                                           ParticipantContext participantContext,
+                                           KinoticSystemService kinoticSystemService,
+                                           @Lazy OrganizationService organizationService,
+                                           @Lazy ApplicationService applicationService) {
         super("kinotic_oidc_configuration", OidcConfiguration.class, esAsyncClient, crudServiceTemplate, participantContext);
+        this.kinoticSystemService = kinoticSystemService;
+        this.organizationService = organizationService;
+        this.applicationService = applicationService;
+    }
+
+    @Override
+    public CompletableFuture<List<OidcConfiguration>> getConfigsForScope(String authScopeType, String authScopeId) {
+        AuthScopeType scope;
+        try {
+            scope = AuthScopeType.valueOf(authScopeType);
+        } catch (IllegalArgumentException e) {
+            return CompletableFuture.failedFuture(
+                    new IllegalArgumentException("Unsupported authScopeType for OIDC lookup: " + authScopeType));
+        }
+
+        return switch (scope) {
+            case SYSTEM -> kinoticSystemService.getOidcConfigurations();
+            case ORGANIZATION -> organizationService.getOidcConfigurations(authScopeId);
+            case APPLICATION -> applicationService.getOidcConfigurations(authScopeId);
+        };
     }
 
     @Override
@@ -53,4 +85,3 @@ public class DefaultOidcConfigurationService extends AbstractCrudService<OidcCon
     }
 
 }
-
